refactor : 사용자 입력 검증부를 Validator 클래스로 분리
